pyproject.toml
[project]
name = "vor-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "distil-whisper-fastrtc>=0.1.0.post1",
    "fastapi[standard]>=0.116.1",
    "faster-whisper>=1.2.0",
    "fastrtc[stt,tts,vad]>=0.0.30",
    "humaware-vad>=0.1.3",
    "librosa>=0.11.0",
    "ollama>=0.5.3",
    "pip>=25.2",
    "torchaudio>=2.8.0",
]
